+++ tests/helpers/test_validate_util.py
@@ -16,6 +16,7 @@ import unittest
 import json
 import sys
 
+from vwo.enums.file_name_enum import FileNameEnum
 from vwo.helpers import validate_util
 from ..data.settings_files import SETTINGS_FILES
 
@@ -53,3 +54,47 @@ class ValidateUtilTest(unittest.TestCase):
         if sys.version_info[0] < 3:
             val = "some_value"
             self.assertIs(False, validate_util.is_valid_unicode(val))
+
+    def test_is_valid_batch_event_settings_events_per_request_passed_below_limit(self):
+        val = {
+            'events_per_request': 0
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)
+
+    def test_is_valid_batch_event_settings_events_per_request_passed_above_limit(self):
+        val = {
+            'events_per_request': 5001
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)
+
+    def test_is_valid_batch_event_settings_events_per_request_passed_with_decimal(self):
+        val = {
+            'events_per_request': 400.2
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)
+
+    def test_is_valid_batch_event_settings_request_time_interval_below_limit(self):
+        val = {
+            'request_time_interval': 0.2
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)
+
+    def test_is_valid_batch_event_settings_missing_values(self):
+        val = {
+            'flush_callback': 1
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)
+
+    def test_is_valid_batch_event_settings_flush_callback_non_callable(self):
+        val = {
+            'events_per_request': 400,
+            'request_time_interval': 10,
+            'flush_callback': 1
+        }
+        result = validate_util.is_valid_batch_event_settings(val, FileNameEnum.Api.Launch)
+        self.assertIs(result, False)