+++ vwo/api/launch.py
@@ -28,7 +28,7 @@ def launch(settings_file, logger=None, user_storage=None, is_development_mode=Fa
     returns a VWO instance.
 
     Args:
-        settings_file (json_string): stringified json representing the settings_file consisting all 
+        settings_file (json_string): stringified json representing the settings_file consisting all
             the campaign related data
         logger (object): an object capable of logging events happening inside the SDK
         user_storage (object): a storage service object capable of doing get and set on
@@ -53,6 +53,7 @@ def launch(settings_file, logger=None, user_storage=None, is_development_mode=Fa
     log_level = kwargs.get("log_level")
     goal_type_to_track = kwargs.get("goal_type_to_track")
     should_track_returning_user = kwargs.get("should_track_returning_user")
+    batch_event_settings = kwargs.get("batch_events")
 
     invalid_log_level = False
     if log_level and not validate_util.is_valid_log_level(log_level):
@@ -73,9 +74,10 @@ def launch(settings_file, logger=None, user_storage=None, is_development_mode=Fa
         or (is_development_mode and type(is_development_mode) is not bool)
         or (goal_type_to_track and not validate_util.is_valid_goal_type(goal_type_to_track))
         or (should_track_returning_user and type(should_track_returning_user) is not bool)
+        or (batch_event_settings and not validate_util.is_valid_batch_event_settings(val=batch_event_settings, file=FILE))
     ):
         logger.log(LogLevelEnum.ERROR, LogMessageEnum.ERROR_MESSAGES.LAUNCH_API_INVALID_PARAMS.format(file=FILE))
         return None
     else:
         logger.log(LogLevelEnum.DEBUG, LogMessageEnum.DEBUG_MESSAGES.VALID_CONFIGURATION.format(file=FILE))
-        return VWO(settings_file, user_storage, is_development_mode, goal_type_to_track, should_track_returning_user)
+        return VWO(settings_file, user_storage, is_development_mode, goal_type_to_track, should_track_returning_user, batch_event_settings)