+++ vwo/enums/log_message_enum.py
@@ -52,6 +52,8 @@ class LogMessageEnum:
             "({file}): [API_NAME] For user_id:{user_id} of campaign:{campaign_key}, whitelisting is skipped"
         )
 
+        BEFORE_FLUSHING = "({file}): Flushing events queue {manually} having {length} events. {timer}, queue summary: {queue_metadata}"
+
     class INFO_MESSAGES:
         """ Classobj encapsulating various INFO messages """
 
@@ -66,6 +68,7 @@ class LogMessageEnum:
         GOT_VARIATION_FOR_USER = "({file}): [API_NAME] user_id:{user_id} for campaign_key:{campaign_key} type: {campaign_type} got variation_name:{variation_name}"
         USER_GOT_NO_VARIATION = "({file}): [API_NAME] user_id:{user_id} for campaign_key:{campaign_key} type: {campaign_type} did not allot any variation"
         IMPRESSION_SUCCESS = "({file}): [API_NAME] Impression event - {end_point} was successfully received by VWO"
+        IMPRESSION_SUCCESS_QUEUE = "({file}): [API_NAME] Impression event - {end_point} was successfully pushed in queue, current queue length : {queue_length}, queue summary: {queue_metadata}"
         MAIN_KEYS_FOR_IMPRESSION = "({file}): [API_NAME] Having main keys: account_id:{account_id} user_id:{user_id} campaignId:{campaign_id} and variation_id:{variation_id}"
         MAIN_KEYS_FOR_PUSH_API = (
             "({file}): [API_NAME] Having main keys: account_id:{account_id} user_id:{user_id} u:{u} and tags:{tags}"
@@ -91,6 +94,10 @@ class LogMessageEnum:
         GOAL_ALREADY_TRACKED = "({file}): [API_NAME] Goal:{goal_identifier} of Campaign:{campaign_key} for User ID:{user_id} has already been tracked earlier. Skipping now. Check docs on how to force tracking for returning visitors."
         SETTINGS_FILE_UPDATED = "({file}): [API_NAME] vwo_sdk_instance is updated with the latest settings_file"
         SETTINGS_FILE_NOT_UPDATED = "({file}): [API_NAME] settings_file fetched are same as earlier fetched settings"
+
+        BULK_IMPRESSION_SUCCESS = "({file}): Impression event - {end_point} was successfully received by VWO having account_id:{account_id}"
+
+        AFTER_FLUSHING = "({file}): Events queue having {length} events has been flushed {manually}, queue summary: {queue_metadata}"
 
     class WARNING_MESSAGES:
         """ Classobj encapsulating various WARNING messages """
@@ -118,6 +125,7 @@ class LogMessageEnum:
         INVALID_CAMPAIGN = "({file}): [API_NAME] Invalid campaign passed to {method} of this file"
         INVALID_USER_ID = "({file}): [API_NAME] Invalid user_id:{user_id} passed to {method} of this file"
         IMPRESSION_FAILED = "({file}): [API_NAME] Impression event could not be sent to VWO - {end_point}"
+        IMPRESSION_FAILED_QUEUE = "({file}): [API_NAME] Impression event could not be pushed in queue - {end_point}"
         CUSTOM_LOGGER_MISCONFIGURED = "({file}): [API_NAME] Custom logger is provided but seems to have misconfigured. {extra_info} Please check the API Docs. Using default logger."
 
         INVALID_API = "({file}): [API_NAME] API is not valid for user ID: {user_id} in campaign ID: {campaign_key} having campaign type: {campaign_type}."
@@ -144,4 +152,16 @@ class LogMessageEnum:
         NO_CAMPAIGN_FOUND = (
             "({file}): [API_NAME] No campaign found for goal_identifier:{goal_identifier}. Please verify from VWO app."
         )
+
+        BULK_NOT_PROCESSED = "({file}): Batch events couldn't be received by VWO. Calling Flush Callback with error and data."
+
+        BATCH_EVENT_LIMIT_EXCEEDED = "({file}): Impression event - {end_point} failed due to exceeding payload size. Parameter events_per_request in batch_events config in launch API has value:{events_per_request} for accountId:{account_id}. Please read the official documentation for knowing the size limits."
         INVALID_SETTINGS_FILE = "({file}): [API_NAME] settings_file fetched is not proper for the account_id: {account_id}."
+
+        EVENT_BATCHING_NOT_OBJECT = "({file}): Batch event settings are not of type object"
+        EVENT_BATCHING_INSUFFICIENT = "({file}): Atleast one from events_per_request and request_time_interval must be provided"
+        EVENTS_PER_REQUEST_INVALID = "({file}): events_per_request should be an integer"
+        REQUEST_TIME_INTERVAL_INVALID = "({file}): request_time_interval should be a number"
+        EVENTS_PER_REQUEST_OUT_OF_BOUNDS = "({file}): events_per_request should be >= {min_value} and <= {max_value}"
+        REQUEST_TIME_INTERVAL_OUT_OF_BOUNDS = "({file}): request_time_interval should be >= {min_value}"
+        FLUSH_CALLBACK_INVALID = "({file}): flush_callback is not callable"